move marked dependency
