+++ src/core/ui.js
@@ -11,7 +11,7 @@
 
 import css from "text!../../assets/ui.css";
 import hyperHTML from "hyperhtml";
-import { markdownToHtml } from "./utils.js";
+import { markdownToHtml } from "./markdown.js";
 import shortcut from "../shortcut.js";
 import { sub } from "./pubsubhub.js";
 export const name = "core/ui";
