+++ src/core/utils.js
@@ -4,32 +4,12 @@
 // As the name implies, this contains a ragtag gang of methods that just don't fit
 // anywhere else.
 import { lang as docLang } from "./l10n.js";
-import marked from "marked";
 import { pub } from "./pubsubhub.js";
 export const name = "core/utils";
-
-marked.setOptions({
-  sanitize: false,
-  gfm: true,
-  headerIds: false,
-});
 
 const spaceOrTab = /^[ |\t]*/;
 const endsWithSpace = /\s+$/gm;
 const dashes = /-/g;
-const gtEntity = /&gt;/gm;
-const ampEntity = /&amp;/gm;
-
-export function markdownToHtml(text) {
-  const normalizedLeftPad = normalizePadding(text);
-  // As markdown is pulled from HTML, > and & are already escaped and
-  // so blockquotes aren't picked up by the parser. This fixes it.
-  const potentialMarkdown = normalizedLeftPad
-    .replace(gtEntity, ">")
-    .replace(ampEntity, "&");
-  const result = marked(potentialMarkdown);
-  return result;
-}
 
 export const ISODate = new Intl.DateTimeFormat(["en-ca-iso8601"], {
   timeZone: "UTC",
